requirements.txt
# Core ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.13.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
boto3>=1.28.0

# Feature Engineering
ta-lib>=0.4.28
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
